test_agent/requirements.txt
prompt_toolkit
paho-mqtt